+++ packages/sdkAbstract/dts/index.d.ts
@@ -1,5 +1,8 @@
 export { UserDeviceInfoImpl } from "./features/contextInformation/userContext";
 export { ErrorStackImpl } from "./features/errorInformation/errorStack";
+export { UserDiyContextImpl, expandTheInformationObjectType } from './features/contextInformation/userDiyContext';
 export { ErrorMessageImpl } from './features/errorInformation/errorMessage';
 export { TriggerLevelImplWithKindType, TriggerLevelImpl, TriggerLevelImplType } from './features/errorInformation/triggerLevel';
 export { MonitoringConfigImpl } from './features/monitoringConfig/monitoringConfig';
+export { HandlingBasicErrors } from './features/helper/handlingBasicErrors';
+export { UserDeviceInfo } from './features/helper/userDeviceInfo';