+++ packages/sdkAbstract/lib/features/contextInformation/userDiyContext.js
@@ -1,0 +1,6 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.UserDiyContextImpl = void 0;
+class UserDiyContextImpl {
+}
+exports.UserDiyContextImpl = UserDiyContextImpl;