+++ packages/sdkAbstract/src/features/errorInformation/triggerLevel.ts
@@ -10,8 +10,7 @@
 
 export type TriggerLevelImplWithKindType = 'stableTrigger' | 'unstableTrigger'
 
-export type TriggerLevelImplType = 'window.error' | 'unhandledrejection' | ''
-
+export type TriggerLevelImplType = 'window.error' | 'unhandledrejection' | 'user-defined' | ''
 
 export abstract class TriggerLevelImpl {