+++ apps/monitoring-platform/CHANGELOG.md
@@ -1,5 +1,11 @@
 # monitoring-platform
 
+## 0.0.2
+
+### Patch Changes
+
+- @memo28.monitoring/service@0.0.3
+
 ## 0.0.1
 
 ### Patch Changes