+++ packages/sdkAbstract/CHANGELOG.md
@@ -1,7 +1,13 @@
 # @memo28.monitoring/sdk-abstract
 
+## 0.0.3
+
+### Patch Changes
+
+- 重构了@memo28.monitoring/sdk-abstract 一些常用类，@memo28.monitoring/sdk-browser 新增自定义抛出事件
+
 ## 0.0.2
 
 ### Patch Changes
 
-- 上传npm仓库使其存在可持续的版本号避免本地install问题
+- 上传 npm 仓库使其存在可持续的版本号避免本地 install 问题